api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.0.0
fastapi==0.110.2
h11==0.14.0
idna==3.7
ipwhois==1.2.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-whois==0.9.3
redis==5.0.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
